requirements.txt
flask==2.2.3
flask-sqlalchemy==3.0.3
flask-cors==3.0.10
python-dotenv==1.0.0
requests==2.28.2
gunicorn==20.1.0
psycopg2-binary==2.9.5